Safari-609.1.11.1/LayoutTests/http/tests/security/module-crossorigin-loads-omit-expected.txt

ALERT: script ran.
This test passes if the module script loads correctly.

PASS
